Adding Checks To Navidrome Connection. Next Release Will Be STABLE

+++ app/src/main/java/com/craftworks/music/providers/navidrome/MarkNavidromeAsPlayed.kt
@@ -15,7 +15,7 @@ fun markNavidromeSongAsPlayed(song: Song){
 
             // Scrobble Percentage
             println("Scrobble Percentage: ${(sliderPos.intValue.toFloat() / SongHelper.currentSong.duration.toFloat()) * 100f}, with sliderPos = ${sliderPos.intValue} | songDuration = ${SongHelper.currentSong.duration} | minPercentage = ${SongHelper.minPercentageScrobble}")
-            if ((sliderPos.intValue.toFloat() / SongHelper.currentSong.duration.toFloat()) * 100f < SongHelper.minPercentageScrobble) return@Thread
+            if ((sliderPos.intValue.toFloat() / SongHelper.currentSong.duration.toFloat()) * 100f < SongHelper.minPercentageScrobble.intValue) return@Thread
 
             println("Scrobble Checks Passed. Contacting Navidrome Server")
